Added confirmation box to delete profile

+++ src/UI/Subsections/UserSettingsFrame.py
@@ -191,7 +191,7 @@ class UserSettingsFrame(AppFrameInterface):
             values=(1, 2, 3, 4, 5, 6, 7, 8, 9, 10)
         )
         self.__pixelSpacing.current(defaultSP.pixelSpacing)
-        self.__pixelSpacing.bind("<<ComboboxSelected>>", self.__user_profile_changed)
+        self.__pixelSpacing.bind("<<ComboboxSelected>>", self.__pixel_spacing_changed)
 
         # hash key label
         self.__hashKeyLabel = ttk.Label(
@@ -295,7 +295,7 @@ class UserSettingsFrame(AppFrameInterface):
         for item in self.__userProfileList:
             if item.name[0][0] == selectedProfileName:
                 self.__selectedUserProfile = item
-        self.__refresh_user_profiles()
+        self.__refresh_user_profiles
 
     def __new_user_profile_pressed(self):
         self._error_if_not_initialized()
@@ -305,15 +305,15 @@ class UserSettingsFrame(AppFrameInterface):
 
     def __delete_user_profile_pressed(self):
         self._error_if_not_initialized()
-        self.__profileService.delete_user_profile(self.__selectedUserProfile.key)
-        self.__refresh_user_profiles
+        ms_bx = tk.messagebox.askquestion("Delete User Profile", "Are you sure you want to delete the User Profile?", icon = 'warning')
+        if ms_bx:
+            self.__profileService.delete_user_profile(self.__selectedUserProfile.key)
+            self.__refresh_user_profiles
 
     def __settings_profile_changed(self, *args):
         self._error_if_not_initialized()
-        selectedProfileName = self.__settingsProfile.get()
-        for item in self.__settingsProfileList:
-            if item.name[0][0] == selectedProfileName:
-                self.__selectedSettingsProfile = item
+        selectedProfilePos = self.__settingsProfile.current()
+        self.__selectedSettingsProfile = self.__settingsProfileList[selectedProfilePos]
         self.__refresh_user_profiles()
 
     def __new_settings_profile_pressed(self):
@@ -326,8 +326,10 @@ class UserSettingsFrame(AppFrameInterface):
 
     def __delete_settings_profile_pressed(self):
         self._error_if_not_initialized()
-        self.__profileService.delete_settings_profile(self.__selectedSettingsProfile.key)
-        self.__refresh_settings_profiles
+        msg_bx = tk.messagebox.askquestion("Delete Settings Profile", "Are you sure you want to delete the Settings Profile?", icon = 'warning')
+        if msg_bx:
+            self.__profileService.delete_settings_profile(self.__selectedSettingsProfile.key)
+            self.__refresh_settings_profiles
 
     def __save_settings_profile_pressed(self):
         self._error_if_not_initialized()
@@ -350,7 +352,7 @@ class UserSettingsFrame(AppFrameInterface):
         self._error_if_not_initialized()
         pass
 
-    def __pixel_spacing_changed(self):
+    def __pixel_spacing_changed(self, *args):
         self._error_if_not_initialized()
         pass
 
@@ -434,8 +436,8 @@ class UserSettingsFrame(AppFrameInterface):
         self.__settingsProfileList = [default]
         if self.__selectedUserProfile != None:
             self.__selectedUserProfile = self.__profileService.get_settings_profiles_for_user(self.__selectedUserProfile)
-            print(self.__selectedUserProfile.settingsProfiles)
             self.__settingsProfileList = self.__selectedUserProfile.settingsProfiles
+            print(self.__settingsProfileList)
             for item in self.__selectedUserProfile.settingsProfiles:
                 self.__settingsProfileNames.append(item.name)
             '''
@@ -444,9 +446,7 @@ class UserSettingsFrame(AppFrameInterface):
             '''
             # TODO update input fields
             print(self.__settingsProfileNames)
-        self.__selectedSettingsProfile = self.__get_default_settings_profile()
-        self.__pixelSpacing["text"] = self.__selectedSettingsProfile.pixelSpacing
-        self.__hashKey["text"] = self.__selectedSettingsProfile.encryptKey
+        #self.__selectedSettingsProfile = self.__get_default_settings_profile()
         self.__settingsProfile['values'] = self.__settingsProfileNames
         # TODO reload dropdown data
